add all

+++ pytpcc/util/rand.py
@@ -30,7 +30,7 @@
 # -----------------------------------------------------------------------
 
 import random
-import nurand
+from . import nurand
 
 SYLLABLES = [ "BAR", "OUGHT", "ABLE", "PRI", "PRES", "ESE", "ANTI", "CALLY", "ATION", "EING" ]
 
@@ -45,8 +45,8 @@ def NURand(a, x, y):
     global nurandVar
     assert x <= y
     if nurandVar is None:
-		setNURand(nurand.makeForLoad())
-    
+        setNURand(nurand.makeForLoad())
+
     if a == 255:
         c = nurandVar.cLast
     elif a == 1023:
@@ -129,7 +129,7 @@ def makeLastName(number):
     """A last name as defined by TPC-C 4.3.2.3. Not actually random."""
     global SYLLABLES
     assert 0 <= number and number <= 999
-    indicies = [ number/100, (number/10)%10, number%10 ]
+    indicies = [ int(number/100), int((number/10)%10), int(number%10) ]
     return "".join(map(lambda x: SYLLABLES[x], indicies))
 ## DEF
